vectorizing_data/requirements.txt
pymilvus==2.4.3
milvus-cli==0.4.3
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
transformers==4.44.2